requirements.txt
fastapi
uvicorn
gunicorn
pydantic
qdrant-client
openai
SQLAlchemy
python-dotenv